diplomeBackend/diplomeBackend/src/main/java/com/example/diplomeBackend/dto/FriendsDTO.java
package com.example.diplomeBackend.dto;

import java.util.List;

public class FriendsDTO {
    private Long id;
    private Long userId;
    private List<Long> friendIds;

    // Getters and Setters
    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public Long getUserId() {
        return userId;
    }

    public void setUserId(Long userId) {
        this.userId = userId;
    }

    public List<Long> getFriendIds() {
        return friendIds;
    }

    public void setFriendIds(List<Long> friendIds) {
        this.friendIds = friendIds;
    }
}
